rename client -> gowandb
