+++ nexus/lib/core/runkeeper.go
@@ -1,24 +1,24 @@
 package main
 
 import (
-	"github.com/wandb/wandb/nexus/pkg/client"
+	"github.com/wandb/wandb/nexus/pkg/gowandb"
 )
 
 type RunKeeper struct {
-	runs map[int]*client.Run
+	runs map[int]*gowandb.Run
 }
 
 func NewRunKeeper() *RunKeeper {
 	return &RunKeeper{
-		runs: make(map[int]*client.Run),
+		runs: make(map[int]*gowandb.Run),
 	}
 }
 
-func (k *RunKeeper) Get(num int) *client.Run {
+func (k *RunKeeper) Get(num int) *gowandb.Run {
 	return k.runs[num]
 }
 
-func (k *RunKeeper) Add(run *client.Run) int {
+func (k *RunKeeper) Add(run *gowandb.Run) int {
 	num := 42
 	k.runs[num] = run
 	return num
