+++ nexus/pkg/gowandb/connection.go
@@ -1,4 +1,4 @@
-package client
+package gowandb
 
 import (
 	"bufio"
