+++ nexus/pkg/gowandb/mailbox.go
@@ -1,4 +1,4 @@
-package client
+package gowandb
 
 import (
 	// "fmt"
