+++ nexus/pkg/gowandb/settings.go
@@ -1,4 +1,4 @@
-package client
+package gowandb
 
 import (
 	"crypto/rand"
